Location updates
Since point was not how we were addressing locations on board I had to
update all our location objects in Player and settlement. In this
process i thought it would be ideal to to have a special location object
for structures since they will be places at an angle. Tests have also
been updated to work with new location changes.