mutation fix for correct numbers and tiletypes on the board.
There is a chance that the bug will not be squashed since the tiletypes and numbers are shuffled, and there are more than 18 in either category and it just so happens that the bottom 18 are the correct acceptable values. In which case, it's not really a bug anymore is it.